Begins basic Voyager navigation implementation
