+++ composeApp/build.gradle.kts
@@ -51,6 +51,9 @@ kotlin {
             implementation(libs.kotlinx.coroutines.core)
             implementation(libs.ktor.core)
             implementation(libs.composeIcons.featherIcons)
+            //The below was needed to finally get the settings to "see" Nav components
+            //implementation("org.jetbrains.androidx.navigation:navigation-compose:2.8.0-alpha01")
+            //implementation("androidx.navigation:navigation-compose:2.4.0-rc01")
         }
 
         commonTest.dependencies {
