+++ composeApp/src/commonMain/kotlin/cherry/company/client_intake_app/App.kt
@@ -11,6 +11,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.unit.dp
+import cafe.adriel.voyager.navigator.Navigator
 import cherry.company.client_intake_app.composables.ShowMoreButton
 import client_intake_app.composeapp.generated.resources.*
 import cherry.company.client_intake_app.theme.AppTheme
@@ -19,7 +20,7 @@ import org.jetbrains.compose.resources.Font
 import org.jetbrains.compose.resources.vectorResource
 import cherry.company.client_intake_app.composables.CheckableRow
 import cherry.company.client_intake_app.composables.DarkModeButton
-import cherry.company.client_intake_app.composables.NewClientForm
+import cherry.company.client_intake_app.composables.NewClientFormScreen
 
 @Composable
 internal fun App() = AppTheme {
@@ -44,7 +45,7 @@ internal fun App() = AppTheme {
 
         DarkModeButton()
 
-        NewClientForm(clientsList)
+        Navigator(NewClientFormScreen(clientsList))
 
         if (clientsList.isNotEmpty()) {
             for (client in clientsList) {
