+++ composeApp/src/commonMain/kotlin/cherry/company/client_intake_app/composables/Navigation.kt
@@ -1,0 +1,8 @@
+package cherry.company.client_intake_app.composables
+
+//@Composable
+//fun Navigation() {
+//    val navController = rememberNavController()
+//    //TODO: Either figure out why 'startDestination' is not being recognized or implement with 'Voyager'
+//    NavHost(navController = navController, startDestination = Screen.CreateClientScreen.route)
+//}
